Pull out wiki data class into its own package
